Refactor distribution tests to not be one huge file
Conflicts:
	dist/jstat.js
	dist/jstat.min.js
	test/distribution/distribution-test.js
