--- test/distribution/distribution-test.js
@@ -1,385 +1,0 @@
-var vows = require('vows');
-var assert = require('assert');
-var suite = vows.describe('jStat.distribution');
-
-require('../env.js');
-
-suite.addBatch({
-  'beta pdf': {
-    'topic': function() {
-      return jStat;
-    },
-    'check pdf calculation': function(jStat) {
-      // Non-log form of the Beta pdf
-      function pdf(x, alpha, beta) {
-        if (x > 1 || x < 0)
-          return 0;
-        return (Math.pow(x, alpha - 1) * Math.pow(1 - x, beta - 1)) /
-            jStat.betafn(alpha, beta);
-      }
-
-      var tol = 0.0000001;
-      var args = [0, 0.1, 0.3, 0.5, 0.7, 0.9, 1];
-      var arg;
-
-      for (var i = 0; i < args.length; i++) {
-        arg = args[i];
-        assert.epsilon(tol, jStat.beta.pdf(arg, 0.1, 0.1), pdf(arg, 0.1, 0.1));
-        assert.epsilon(tol, jStat.beta.pdf(arg, 1, 1), pdf(arg, 1, 1));
-        assert.epsilon(tol, jStat.beta.pdf(arg, 10, 50), pdf(arg, 10, 50));
-
-        // Show that the log form of the pdf performs better for
-        // large parameter values.
-        assert(!isNaN(jStat.beta.pdf(arg, 1000, 5000)),
-               'New Beta pdf is NaN for large parameter values.');
-        assert(isNaN(pdf(arg, 1000, 5000)),
-               'Old Beta pdf is not NaN for large parameter values.');
-      }
-
-      assert.equal(jStat.beta.pdf(0, 1, 4), 4);
-      assert.equal(jStat.beta.pdf(1, 4, 1), 4);
-    }
-  },
-
-  'gamma pdf': {
-    'topic': function() {
-      return jStat;
-    },
-    'check instance and static pdf method': function (jStat) {
-      var shape = 5;
-      var scale = 1;
-      var gamma = jStat.gamma(shape, scale);
-      var xValues = [-1, 0, 1];
-      var x;
-      for (var i = 0; i < xValues.length; i++) {
-        x = xValues[i];
-        pStatic = jStat.gamma.pdf(x, shape, scale);
-        pInstance = gamma.pdf(x);
-        if (isNaN(pStatic)) {
-          assert(isNaN(pInstance));
-        } else {
-          assert(pStatic === pInstance,
-                 'Gamma pdf evaluated at ' +
-                   x + ' should be equal for instance and static methods.');
-        }
-      }
-    }
-  },
-
-  'uniform pdf': {
-    'topic': function() {
-      return jStat;
-    },
-    'check pdf calculation': function(jStat) {
-      var tol = 0.0000001;
-      assert.epsilon(tol, jStat.uniform.pdf(10.5, 10, 11), 1.0);
-      assert.epsilon(tol, jStat.uniform.pdf(6, 0, 10), 0.1);
-    },
-    'check cdf calculation': function(jStat) {
-      var tol = 0.0000001;
-      assert.epsilon(tol, jStat.uniform.cdf(7, 0, 10), 0.7);
-      assert.epsilon(tol, jStat.uniform.cdf(10.5, 10, 11), 0.5);
-    },
-    'check inv calculation': function(jStat) {
-      var tol = 0.0000001;
-      assert.epsilon(tol, jStat.uniform.inv(0.5, 10, 11), 10.5);
-      assert.epsilon(tol, jStat.uniform.inv(0.7, 0, 10), 7.0);
-    }
-  },
-
-  'binomial pdf': {
-    'topic': function() {
-      return jStat;
-    },
-    //checked against R's dbinom(k, n, p)
-    'check pdf calculation': function(jStat) {
-      var tol = 0.0000001;
-      assert.epsilon(tol, jStat.binomial.pdf(10, 25, 0.5), 0.09741664);
-      assert.epsilon(tol, jStat.binomial.pdf(50, 1000, 0.05), 0.05778798);
-    },
-    //Checked against r's pbinom(k, n, p)
-    'check cdf calculation': function(jStat) {
-      var tol = 0.0000001;
-      assert.epsilon(tol, jStat.binomial.cdf(10, 25, 0.5), 0.2121781);
-      assert.epsilon(tol, jStat.binomial.cdf(50, 1000, 0.05), 0.537529);
-    }
-  },
-
-  'central F distribution': {
-    'topic': function() {
-      return jStat;
-    },
-    // Check against R's df(x, df1, df2)
-    'check pdf calculation': function(jStat) {
-      var tol = 0.0000001;
-
-      var zeroth = jStat.centralF.pdf(0.2, 1, 3);
-      assert.isNumber(zeroth);
-      assert.epsilon(tol, zeroth, 0.722349);
-
-      var first = jStat.centralF.pdf(1, 100, 100);
-      assert.isNumber(first);
-      assert.epsilon(tol, first, 1.989731);
-
-      var second = jStat.centralF.pdf(2.5, 50, 200);
-      assert.isNumber(second);
-      assert.epsilon(tol, second, 0.00003610325);
-
-      var third = jStat.centralF.pdf(0.8, 2, 10);
-      assert.isNumber(third);
-      assert.epsilon(tol, third, 0.4104423);
-
-      var value = jStat.centralF.pdf(0.4, 3, 10);
-      assert.isNumber(value);
-      assert.epsilon(tol, value, 0.6733766);
-    }
-  },
-
-  'studentt pdf': {
-    'topic': function() {
-      return jStat;
-    },
-    //Checked against R dt(x,df)
-    'check pdf calculation': function(jStat) {
-      var tol = 0.0000001;
-      var x = 0.5;
-      var df = 40;
-      assert.epsilon(tol, jStat.studentt.pdf(x, df), 0.3489195);
-    },
-    //Checked against R's pt(x, df)
-    'check cdf calculation': function(jStat) {
-      var tol = 0.0000001;
-      var x = 0.5;
-      var df = 40;
-      assert.epsilon(tol, jStat.studentt.cdf(x, df), 0.6900926);
-    },
-    //Checked against R's qt(x, df)
-    'check inv calculation': function(jStat) {
-      var tol = 0.0000001;
-      assert.epsilon(tol, jStat.studentt.inv(0.5, 40), 0);
-      assert.epsilon(tol, jStat.studentt.inv(0.7, 20), 0.5328628);
-      assert.epsilon(tol, jStat.studentt.inv(0.2, 10), -0.8790578);
-    }
-  },
-});
-
-suite.addBatch({
-  'negbin pdf': {
-    'topic': function() {
-      return jStat;
-    },
-    'check pdf calculation': function(jStat) {
-      var k = 10; // number of failures
-      var r = 5; // threshold number of successes
-      var p = 0.25; // probability of a success
-      assert(jStat.negbin.pdf(k, r, p), 0.05504866037517786);
-    }
-  },
-
-  'negbin cdf': {
-    'topic': function() {
-      return jStat;
-    },
-    'check pdf calculation': function(jStat) {
-      var k = 10; // number of failures
-      var r = 5; // threshold number of successes
-      var p = 0.25; // probability of a success
-      assert(jStat.negbin.cdf(k, r, p), 0.3135140584781766);
-    }
-  },
-});
-
-suite.addBatch({
-  'hypergeometric pdf': {
-    'topic': function() {
-      return jStat;
-    },
-    'check pdf calculation': function(jStat) {
-      var tol = 0.0000001;
-      // How many 1s were obtained by sampling?
-      var successes = [10, 16];
-      // How big was the source population?
-      var population = [100, 3589];
-      // How many 1s were in it?
-      var available = [20, 16];
-      // How big a sample was taken?
-      var draws = [15, 2290];
-      // What was the probability of exactly this many 1s?
-      // Obtained from the calculator at
-      // <http://www.geneprof.org/GeneProf/tools/hypergeometric.jsp>
-      var answers = [0.000017532028090435493, 0.0007404996809672229];
-
-      for (var i = 0; i < answers.length; i++) {
-        // See if we get the right answer for each calculation.
-        var calculated = jStat.hypgeom.pdf(successes[i],
-                                           population[i],
-                                           available[i],
-                                           draws[i]);
-        // None of the answers should be NaN
-        assert(!isNaN(calculated));
-        // And they should all match
-        assert.epsilon(tol, calculated, answers[i]);
-      }
-    }
-  }
-});
-
-suite.addBatch({
-  'hypergeometric pdf': {
-    'topic': function() {
-      return jStat;
-    },
-    'check pdf agreement': function(jStat) {
-      var tol = 0.0000001;
-
-      // Define an easy pdf
-      var easyPDF = function pdf(k, N, m, n) {
-        if (k !== (k | 0))
-          return false;
-        if (k < 0)
-          return 0;
-        return jStat.combination(m, k) * jStat.combination(N - m , n - k) /
-            jStat.combination(N, n);
-      }
-
-      // Now test the easy PDF agaisnt the real one.
-      for (var N = 0; N < 10; N++) {
-        // For several population sizes
-        for (var m = 0; m < N; m++) {
-          // For all possible success counts
-          for (var n = 0; n < N; n++) {
-            // For all possible sample sizes
-            for (var x = 0; x < n; x++) {
-              // For all subset sizes of the sampled set
-              // Get the probability be each method
-              var probEasy = easyPDF(x, N, m, n);
-
-              if (!isNaN(probEasy)) {
-                // The easy PDF worked for this situation. Compare it to the
-                // real one.
-
-                var probReal = jStat.hypgeom.pdf(x, N, m, n);
-
-                assert(!isNaN(probReal), 'Hypergeometric PDF returned NaN');
-                assert.epsilon(tol,
-                    probReal,
-                    probEasy,
-                    'Hypergeometric PDF didn\'t match naive implementation');
-              }
-            }
-          }
-        }
-      }
-    },
-
-    'check cdf agreement': function(jStat) {
-      var tol = 0.0000001;
-
-      // Define an easy pdf
-      function easyPDF(k, N, m, n) {
-        if (k !== (k | 0))
-          return false;
-        if (k < 0)
-          return 0;
-        return jStat.combination(m, k) * jStat.combination(N - m , n - k) /
-            jStat.combination(N, n);
-      }
-
-      // Define a summation cdf based on it
-      function easyCDF(x, N, m, n) {
-        // Sum up the total probability of every number of successes <= x, and
-        // put it here.
-        var sum = 0;
-        for (var i = 0; i <= x; i++)
-          sum += easyPDF(i, N, m, n);
-        return sum;
-      }
-
-      // Now test the easy CDF agaisnt the real one.
-      for (var N = 0; N < 10; N++) {
-        // For several population sizes
-        for (var m = 0; m < N; m++) {
-          // For all possible success counts
-          for (var n = 0; n < N; n++) {
-            // For all possible sample sizes
-            for (var x = 0; x < n; x++) {
-              // For all subset sizes of the sampled set
-              // Get the probability be each method
-              var probEasy = easyCDF(x, N, m, n);
-
-              if (!isNaN(probEasy)) {
-                // The easy CDF worked for this situation. Compare it to the
-                // real one.
-                var probReal = jStat.hypgeom.cdf(x, N, m, n);
-                assert(!isNaN(probReal), 'Hypergeometric CDF returned NaN');
-                assert.epsilon(tol,
-                    probReal,
-                    probEasy,
-                    'Hypergeometric CDF didn\'t match naive implementation');
-              }
-            }
-          }
-        }
-      }
-    },
-
-    'check pdf calculation': function(jStat) {
-      var tol = 0.0000001;
-      // How many 1s were obtained by sampling?
-      var successes = [ 10, 16, 252 ];
-      // How big was the source population?
-      var population = [ 100, 3589, 3589 ];
-      // How many 1s were in it?
-      var available = [ 20, 16, 252 ];
-      // How big a sample was taken?
-      var draws = [ 15, 2290, 252 ];
-      // What was the probability of exactly this many 1s?
-      // Obtained from the calculator at
-      // <http://www.geneprof.org/GeneProf/tools/hypergeometric.jsp>
-      var answers = [ 0.000017532028090435493, 0.0007404996809672229, 0 ];
-
-      for (var i = 0; i < answers.length; i++) {
-        // See if we get the right answer for each calculation.
-        var calculated = jStat.hypgeom.pdf(successes[i], population[i],
-                                           available[i], draws[i]);
-
-        // None of the answers should be NaN
-        assert(!isNaN(calculated), "Hypergeometric PDF returned NaN");
-
-        // And they should all match
-        assert.epsilon(tol, calculated, answers[i],
-          "Hypergeometric PDF returned incorrect answer");
-      }
-    },
-    'check cdf calculation': function(jStat) {
-      var tol = 0.0000001;
-      // How many 1s were obtained by sampling?
-      var successes = [ 10, 16, 10 ];
-      // How big was the source population?
-      var population = [ 100, 3589, 3589 ];
-      // How many 1s were in it?
-      var available = [ 20, 16, 16 ];
-      // How big a sample was taken?
-      var draws = [ 15, 2290, 2290 ];
-      // What was the probability of this many 1s or fewer?
-      // Obtained from the calculator at
-      // <http://www.geneprof.org/GeneProf/tools/hypergeometric.jsp>
-      var answers = [ 0.9999989096, 1, 0.55047323949 ];
-
-      for (var i = 0; i < answers.length; i++) {
-        // See if we get the right answer for each calculation.
-        var calculated = jStat.hypgeom.cdf(successes[i], population[i],
-                                           available[i], draws[i]);
-
-        // None of the answers should be NaN
-        assert(!isNaN(calculated), "Hypergeometric CDF returned NaN");
-
-        // And they should all match
-        assert.epsilon(tol, calculated, answers[i],
-          "Hypergeometric CDF returned incorrect answer");
-      }
-    }
-  }
-});
-
-suite.export(module);
